LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/choice-of-error-3-expected.txt

CONSOLE MESSAGE: [object Object]
CONSOLE MESSAGE: [object Object]

PASS Evaluation errors are cached in intermediate module scripts
